# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ISRO API",
    "version": "v1",
    "description": "ISRO data API with JWT auth."
  },
  "servers": [
    { "url": "/" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/spacecraft": {
      "get": {"summary": "List spacecraft missions", "parameters": [
        {"in":"query","name":"q","schema":{"type":"string"}},
        {"in":"query","name":"page","schema":{"type":"integer"}},
        {"in":"query","name":"limit","schema":{"type":"integer"}},
        {"in":"query","name":"sort","schema":{"type":"string"}}
      ],"responses":{"200":{"description":"OK"}}}
    },
    "/api/launches": {
      "get": {"summary": "List launches (Mongo)", "parameters": [
        {"in":"query","name":"q","schema":{"type":"string"}},
        {"in":"query","name":"page","schema":{"type":"integer"}},
        {"in":"query","name":"limit","schema":{"type":"integer"}},
        {"in":"query","name":"sort","schema":{"type":"string","example":"launch_date:desc"}}
      ],"responses":{"200":{"description":"OK"}}}
    },
    "/api/timeline": {"get": {"summary": "Timeline", "responses": {"200": {"description": "OK"}}}},
    "/api/upcoming": {"get": {"summary": "Upcoming missions", "responses": {"200": {"description": "OK"}}}},
    "/api/details": {"get": {"summary": "Mission details", "responses": {"200": {"description": "OK"}}}},
    "/api/search": {"get": {"summary": "Cross-collection search", "parameters": [
      {"in":"query","name":"q","required":true,"schema":{"type":"string"}},
      {"in":"query","name":"page","schema":{"type":"integer"}},
      {"in":"query","name":"limit","schema":{"type":"integer"}}
    ], "responses": {"200": {"description": "OK"}}}},
    "/api/analytics/launches": {"get": {"summary": "Launch analytics (grouped)", "responses": {"200": {"description": "OK"}}}},
    "/api/datasets": {"get": {"summary": "Datasets catalog (Mongo)", "parameters": [
      {"in":"query","name":"q","schema":{"type":"string"}},
      {"in":"query","name":"page","schema":{"type":"integer"}},
      {"in":"query","name":"limit","schema":{"type":"integer"}}
    ], "responses": {"200": {"description": "OK"}}}},
    "/enquire": {"get": {"summary": "AI-assisted enquiry (JWT required)", "parameters": [
      {"in":"query","name":"q","required":true,"schema":{"type":"string"}}
  ], "responses": {"200": {"description": "OK"}}}},
  "/ai/summarize": {
      "post": {
        "security": [{"bearerAuth": []}],
        "summary": "Summarize text or API data",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"text": {"type": "string"}, "endpoint": {"type": "string"}, "params": {"type": "object"}}}}}},
        "responses": {"200": {"description": "OK"}}
      }
    },
  "/ai/compare": {
      "post": {
        "security": [{"bearerAuth": []}],
        "summary": "Compare two items by query",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"type": {"type": "string"}, "left": {"type": "string"}, "right": {"type": "string"}}, "required": ["left","right"]}}}},
        "responses": {"200": {"description": "OK"}}
      }
    },
  "/ai/extract": {
      "post": {
        "security": [{"bearerAuth": []}],
        "summary": "Extract fields from text",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"text": {"type": "string"}, "schema": {"type": "object"}}, "required": ["text"]}}}},
        "responses": {"200": {"description": "OK"}}
      }
    },
    "/ai/rewrite": {
      "post": {
        "security": [{"bearerAuth": []}],
        "summary": "Rewrite text with tone and length",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"text": {"type": "string"}, "tone": {"type": "string"}, "length": {"type": "string"}}, "required": ["text"]}}}},
        "responses": {"200": {"description": "OK"}}
      }
    },
    "/ai/classify": {
      "post": {
        "security": [{"bearerAuth": []}],
        "summary": "Classify text into labels",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"text": {"type": "string"}, "labels": {"type": "array", "items": {"type": "string"}}}, "required": ["text"]}}}},
        "responses": {"200": {"description": "OK"}}
      }
    },
    "/ai/qna": {
      "post": {
        "security": [{"bearerAuth": []}],
        "summary": "QnA over API data",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"question": {"type": "string"}, "endpoint": {"type": "string"}, "params": {"type": "object"}}, "required": ["question"]}}}},
        "responses": {"200": {"description": "OK"}}
      }
    },
    "/ai/usage/today": {"get": {"security": [{"bearerAuth": []}], "summary": "Today usage and cost (estimated)", "responses": {"200": {"description": "OK"}}}},
    "/ai/usage": {"get": {"security": [{"bearerAuth": []}], "summary": "Usage by day and model", "parameters": [{"in":"query","name":"days","schema":{"type":"integer","default":7}}], "responses": {"200": {"description": "OK"}}}},
    "/api/news": {"get": {"security": [{"bearerAuth": []}], "summary": "Press releases", "responses": {"200": {"description": "OK"}}}},
    "/api/launch-vehicle-specs": {"get": {"security": [{"bearerAuth": []}], "summary": "Launch vehicle page snapshots", "responses": {"200": {"description": "OK"}}}}
  }
}